fix: nav ui kit composition tidy up and pattern code blocks
